gio  7 giu 2018, 02.25.03, CEST

+++ include/sys/_sigset.h
@@ -38,6 +38,6 @@
 #ifndef _SYS__SIGSET_H_
 #define	_SYS__SIGSET_H_
 
-typedef unsigned long __sigset_t;
+typedef unsigned long long __sigset_t;
 
 #endif /* !_SYS__SIGSET_H_ */
